requirements.txt
pandas
numpy
scipy
ipykernel
pyyaml